change int to double

+++ Luzyce.Core/Models/ProductionPlan/UpdateInfoInProductionPlanPosition.cs
@@ -5,7 +5,7 @@ public class UpdateInfoInProductionPlanPosition
     public int Id { get; set; }
     public int? GetHeadsOfMetallurgicalTeamsId { get; set; }
     public int? NumberOfHours { get; set; }
-    public int? WeightNetto { get; set; }
-    public int? WeightBrutto { get; set; }
+    public double? WeightNetto { get; set; }
+    public double? WeightBrutto { get; set; }
     public int? QuantityPerChange { get; set; }
 }
